requirements.txt
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=1.0.0
pycaret>=3.0.0
python-dotenv>=0.19.0
matplotlib>=3.4.0
seaborn>=0.11.0
joblib>=1.0.0
ipywidgets>=7.6.5 